[FLINK-37213][table-runtime] Improve performance of unbounded OVER aggregations
Instead of sorting all of the records based on the row time
explicilty use timers to achieve the same thing.

This version vs the previous one register a timer for each
record, as opposed to just one timer per key. However since
we are using RocksDB for timers, this is a minor problem.
In exchange, we:
 - don't have to iterate over all of the state for each timer
 - we are firing timers only when needed, vs for each watermark
   for each key. For example if watermarks are fire every 200ms
   and for a given key, we have only one record that should be
   fired 20s into the future, the previous version would be
   firing a timer for that key for each watermark unnecessarily
   without doing any work.